requirements.txt
streamlit
pillow
numpy
opencv-python
pytesseract
